Optimizations in StokesBuildManager
Changed index order in build_system_of_equations() to align with Julia
column-major memory layout.

Type of attribute hshift_to_vxR changed to Int64 (from Float64).

+++ src/stokes_continuity_solver/solver/build_system/utils/BuildStructs.jl
@@ -67,7 +67,7 @@ struct GridData
     ynum::Int64
     xstpavg::Float64
     ystpavg::Float64
-    hshift_to_vxR::Float64
+    hshift_to_vxR::Int64
     xstp::Vector{Float64}
     ystp::Vector{Float64}
     xstpc::Vector{Float64}
